+++ myers_diff.gemspec
@@ -1,12 +1,12 @@
 Gem::Specification.new do |s|
   s.name = 'myers_diff'
-  s.version = '0.1.2'
-  s.date = '2020-02-16'
+  s.version = '1.0.0'
+  s.date = '2020-05-22'
   s.summary = 'Implementation of Myers 1986 text diff algorithm'
   s.description = 'Implementation of Myers 1986 text diff algorithm'
   s.authors = ['Alex Tsui']
   s.email = 'alextsui@pm.me'
-  s.files = ['lib/myers_diff.rb']
+  s.files = Dir.glob('lib/**/*')
   s.homepage = 'https://github.com/alextsui05/myers_diff'
   s.license = 'MIT'
 end
